requirements.txt
asgiref==3.5.0
certifi==2021.10.8
Django==4.0.2
django-model-utils==4.2.0
djangorestframework==3.13.1
mysqlclient @ file:///opt/concourse/worker/volumes/live/3e2c08bc-c6d4-44b9-6985-8c1059f99ec2/volume/mysqlclient_1609786255460/work
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
